refactor: refactor DTO packages
