--- src/main/java/fun/trackmoney/auth/dto/internal/LoginFailure.java
@@ -1,3 +1,0 @@
-package fun.trackmoney.auth.dto.internal;
-
-public record LoginFailure(AuthError error) implements LoginResult { }
