--- src/main/java/fun/trackmoney/auth/dto/internal/UserRegisterFailure.java
@@ -1,4 +1,0 @@
-package fun.trackmoney.auth.dto.internal;
-
-public record UserRegisterFailure(AuthError errorList) implements UserRegisterResult {
-}
