+++ src/main/java/fun/trackmoney/auth/dto/internal/login/LoginFailure.java
@@ -1,0 +1,5 @@
+package fun.trackmoney.auth.dto.internal.login;
+
+import fun.trackmoney.auth.dto.internal.AuthError;
+
+public record LoginFailure(AuthError error) implements LoginResult { }
