+++ src/main/java/fun/trackmoney/auth/dto/internal/login/LoginResult.java
@@ -1,3 +1,3 @@
-package fun.trackmoney.auth.dto.internal;
+package fun.trackmoney.auth.dto.internal.login;
 
 public sealed interface LoginResult permits LoginSuccess, LoginFailure { }
