+++ src/main/java/fun/trackmoney/auth/dto/internal/register/UserRegisterFailure.java
@@ -1,0 +1,6 @@
+package fun.trackmoney.auth.dto.internal.register;
+
+import fun.trackmoney.auth.dto.internal.AuthError;
+
+public record UserRegisterFailure(AuthError errorList) implements UserRegisterResult {
+}
