+++ src/main/java/fun/trackmoney/auth/dto/internal/register/UserRegisterResult.java
@@ -1,4 +1,4 @@
-package fun.trackmoney.auth.dto.internal;
+package fun.trackmoney.auth.dto.internal.register;
 
 public sealed interface UserRegisterResult permits UserRegisterSuccess, UserRegisterFailure {
 }
